Modify GstAudioDevice to be configurable as live/non-live source
GStreamer source elements have some particularities that must be
followed and one of the main consequences is a increased latency time.
WebKitWebAudioSrc didn't operate as a live source until live-input
support has been added. After that is desirable Nix::AudioDevice
implementation works in both modes, live and non-live, depending on the
construction parameters. For example, when 'numberOfInputChannels are 0,
doesn't make sense it to operate in live mode.

This patch adds this feature and besides that it adds some new
properties to WebKitWebAudioSrc regarding number of channels and
removing some hardcoded logic to use this new properties.

Signed-off-by: Nick Diego Yamane <nick.yamane@openbossa.org>